Deprecate current_room and reyna_room attributes in Game
This involves a linear search for the characters' location in each function call.
While current impact is low due to small number of rooms, code should later
be refactored to not require access to current room (or otherwise to pass the room
instance around if necessary).

This paves the groundwork for abilities to modify map data more easily.